Seperate the entityComponents to a seperate header Create the Structure of Arrays in entity manager
