+++ Light Vox Engine/Light Vox Engine.vcxproj
@@ -236,6 +236,7 @@ for /r "$(SolutionDir)_Binary" %%x in (*.cso) do move "%%x" "$(SolutionDir)_Bina
     <ClInclude Include="d3dx12.h" />
     <ClInclude Include="DXSampleHelper.h" />
     <ClInclude Include="Engine.h" />
+    <ClInclude Include="EntityComponents.h" />
     <ClInclude Include="EntityManager.h" />
     <ClInclude Include="FrameResource.h" />
     <ClInclude Include="GraphicsCore.h" />
