+++ Light Vox Engine/Light Vox Engine.vcxproj.filters
@@ -73,6 +73,9 @@
     <ClInclude Include="GameTime.h">
       <Filter>Source Files\Utils</Filter>
     </ClInclude>
+    <ClInclude Include="EntityComponents.h">
+      <Filter>Source Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <FxCompile Include="Shaders\ps_basic.hlsl">
